DevOps: Fix the continuous-deployment workflow
The workflow was broken as the `check_version.py` file was expecting the
tag as the last argument to its invocation, but was really receiving the
``github.ref`` environment variable which includes the `refs/tags/`
prefix. The result was that the check for the tag matching the package
version would always fail.

The workflow is replaced by one that is tried-and-true in many other
repositories of the `aiidateam` organization.