POST `/processes`: Fix deserialization of nested inputs
The function that deserialized the input dictionary, meaning converting
the UUID of any key that ends with `.uuid` into the corresponding node
by loading it from the database, was not recursing into any nested
dictionaries. This means that the endpoint was broken for processes with
nested input namespaces. This is fixed by making the deserialization
function recursive.