DevOps: Add AiiDA deprecation warnings
The warnings are enabled by setting the `AIIDA_WARN_v3` environment
variable to `True` in the CI workflow.

Most of the warnings are addressed, but a few remain that cannot be
trivially fixed:

 * The `id` property on the `aiida.orm.Entity` class is deprecated. This
   is being called because the ORM proxy model declares `id` as the
   primary key field. This is correct for the database models of the
   `PsqlDosBackend` in `aiida-core`, but the front-end ORM is using the
   `pk` property instead.
 * The `attributes`, `extras` and `repository_metadata` of the proxy
   models are declared "top-level" and so `pydantic` will set them on
   the ORM instance using that property, however, these properties have
   been deprecated in `aiida-core` and should be set through the
   `base.attributes.set_many`, `base.extras.set_many` and the
   `base.repository.metadata` properties instead.

With the current setup using `pydantic` it is not clear how these
changes can be addressed and whether the manner in which field values
are being set on the underlying ORM model instance can be customized.