POST `/nodes`: Remove custom entry point group (#50)
The POST `/nodes` end point was relying on a custom entry point group
`aiida.rest.post` which registered the `aiida_restapi.models:Node_Post`
class under a number of entry points. The entry point names correspond
to the `node_type` attributes that are automatically generated and set
on instance of AiiDA `Node` ORM instances.

Not only were the entry points essentially pointless, since they all
load the same class, which can simply be imported since it is a class
provided by `aiida-restapi` itself, it also restricted to what `Node`
types could be constructed. If a particular plugin was not present in
the `aiida.rest.post` entry point group, the POST request would return a
404 error.

Instead, we should simply reuse the unique naming scheme that already
exists in `aiida-core` which is to use existing entry point names for
`Node` plugins, which are provided in the `aiida.data` entry point
group. The `aiida.rest.post` entry point group is removed and the
endpoint now requires the `entry_point` field instead of the
`node_type`. The `entry_point` is loaded using tools from `aiida-core`
and a 404 is returned if the entry point does not exist on the server.

If the entry point can be loaded, the corresponding class is
instantiated. AiiDA's ORM will automatically set the correct `node_type`
through this manner. This now means in principle that any entry point
that is registered on the server can be constructed through the end
point.