Add `nodes/{node_id}/download` endpoint to REST API (#74)
We implement the download of nodes that implement the export
functionality as in aiida-core. We use StreamingResponse to send the
data in byte chunks of size 1024.

The tests require the client to be asynchronously to prevent I/O
operations on closed file handlers.

We require a numpy dependency because we need a node type that
implemented the export functionality. The only node type that is not
bounded to the domain of material science is ArrayData.