Upgrade documentation for migration to aiida-core 2.6 (#82)
Remove usage of `load_documentation_profile` in docs. With
aiida-core/pull/6226 the usage of load_documentation_profile is not
anymore required as the Django backend is not anymore supported and can
therefore be removed.

graphql.utils has been moved to graphql.utilities and usage of
`print_schema` has been adapted.